microseconds in metrics

+++ resources/assets/js/components/Charts/BaseCharts/Line.js
@@ -34,7 +34,7 @@ export default Vue.extend({
                             },
                             beforeBuildTicks: function (scale) {
                                 var max = _.max(scale.chart.data.datasets[0].data);
-                                scale.max = parseInt(max) + parseInt(max * 0.25);
+                                scale.max = parseFloat(max) + parseInt(max * 0.25);
                             },
                         }
                     ],
